+++ src/features/problems/index.ts
@@ -1,4 +1,4 @@
 export * from "./actions";
 export * from "./components";
 export * from "./queries";
-export * from "./types";
+export * from "./type";
