+++ src/features/rankings/index.ts
@@ -1,3 +1,3 @@
 export * from "./components";
 export * from "./queries";
-export * from "./types";
+export * from "./type";
